rename Request back to Requests to avoid type conflicts and make naming consistent with other integrations (ZIOs, Routes)
