+++ kyo-bench/src/main/scala/kyo/bench/HttpClientBench.scala
@@ -35,7 +35,7 @@ class HttpClientBench extends Bench.ForkOnly("pong"):
     override def kyoBenchFiber() =
         import kyo.*
 
-        Abort.run(Request(_.get(kyoUrl))).map(_.getOrThrow)
+        Abort.run(Requests(_.get(kyoUrl))).map(_.getOrThrow)
     end kyoBenchFiber
 
     val zioUrl =
