+++ kyo-bench/src/main/scala/kyo/bench/HttpClientContentionBench.scala
@@ -37,7 +37,7 @@ class HttpClientContentionBench
     override def kyoBenchFiber() =
         import kyo.*
 
-        Async.parallel(Seq.fill(concurrency)(Abort.run(Request(_.get(kyoUrl))).map(_.getOrThrow)))
+        Async.parallel(Seq.fill(concurrency)(Abort.run(Requests(_.get(kyoUrl))).map(_.getOrThrow)))
     end kyoBenchFiber
 
     val zioUrl =
