+++ kyo-sttp/jvm/src/test/scala/kyo/RequestsLiveTest.scala
@@ -3,27 +3,27 @@ package kyo
 import scala.util.*
 import sttp.client3.*
 
-class RequestLiveTest extends Test:
+class RequestsLiveTest extends Test:
 
     "requests" - {
         "live" - {
             "success" in run {
                 for
                     port <- startTestServer("/ping", Success("pong"))
-                    r    <- Request(_.get(uri"http://localhost:$port/ping"))
+                    r    <- Requests(_.get(uri"http://localhost:$port/ping"))
                 yield assert(r == "pong")
             }
             "failure" in run {
                 for
                     port <- startTestServer("/ping", Failure(new Exception))
-                    r    <- Abort.run(Request(_.get(uri"http://localhost:$port/ping")))
+                    r    <- Abort.run(Requests(_.get(uri"http://localhost:$port/ping")))
                 yield assert(r.isFail)
             }
             "race" in run {
                 val n = 1000
                 for
                     port <- startTestServer("/ping", Success("pong"))
-                    r    <- Async.race(Seq.fill(n)(Request(_.get(uri"http://localhost:$port/ping"))))
+                    r    <- Async.race(Seq.fill(n)(Requests(_.get(uri"http://localhost:$port/ping"))))
                 yield assert(r == "pong")
                 end for
             }
@@ -66,4 +66,4 @@ class RequestLiveTest extends Test:
             Resource.ensure(server.stop(0))
                 .andThen(IO(server.getAddress.getPort()))
         }
-end RequestLiveTest
+end RequestsLiveTest
